QRcode/pyproject.toml
[project]
name = "qrcode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypng>=0.20220715.0",
    "pyqrcode>=1.2.1",
]
